refactor: use jotai for currentCamera
