+++ web/src/beta/features/Editor/Map/InspectorPanel/SceneSettings/index.tsx
@@ -3,7 +3,6 @@ import { useMemo } from "react";
 import PropertyItem from "@reearth/beta/components/fields/Property/PropertyItem";
 import { filterVisibleItems } from "@reearth/beta/components/fields/utils";
 import SidePanelSectionField from "@reearth/beta/components/SidePanelSectionField";
-import type { Camera } from "@reearth/beta/utils/value";
 import type { FlyTo } from "@reearth/core";
 import type { Item } from "@reearth/services/api/propertyApi/utils";
 import { useT } from "@reearth/services/i18n";
@@ -12,11 +11,10 @@ import { styled } from "@reearth/services/theme";
 type Props = {
   propertyId: string;
   propertyItems?: Item[];
-  currentCamera?: Camera;
   onFlyTo?: FlyTo;
 };
 
-const Settings: React.FC<Props> = ({ propertyId, propertyItems, currentCamera, onFlyTo }) => {
+const Settings: React.FC<Props> = ({ propertyId, propertyItems, onFlyTo }) => {
   const t = useT();
   const visibleItems = useMemo(() => filterVisibleItems(propertyItems), [propertyItems]);
 
@@ -24,13 +22,7 @@ const Settings: React.FC<Props> = ({ propertyId, propertyItems, currentCamera, o
     <Wrapper>
       {visibleItems?.map((i, idx) => (
         <SidePanelSectionField title={i.title ?? t("Settings")} key={idx}>
-          <PropertyItem
-            key={i.id}
-            propertyId={propertyId}
-            item={i}
-            currentCamera={currentCamera}
-            onFlyTo={onFlyTo}
-          />
+          <PropertyItem key={i.id} propertyId={propertyId} item={i} onFlyTo={onFlyTo} />
         </SidePanelSectionField>
       ))}
     </Wrapper>
