+++ web/src/beta/features/Visualizer/StoryPanel/Block/builtin/Camera/index.tsx
@@ -9,13 +9,7 @@ import { type CameraBlock as CameraBlockType } from "./Editor";
 
 export type Props = BlockProps<StoryBlock>;
 
-const CameraBlock: React.FC<Props> = ({
-  block,
-  isSelected,
-  currentCamera,
-  onPropertyItemAdd,
-  ...props
-}) => {
+const CameraBlock: React.FC<Props> = ({ block, isSelected, onPropertyItemAdd, ...props }) => {
   const cameraButtons = useMemo(
     () => (block?.property?.default ?? []) as CameraBlockType[],
     [block?.property?.default],
@@ -40,7 +34,6 @@ const CameraBlock: React.FC<Props> = ({
       onPropertyItemAdd={onPropertyItemAdd}
       {...props}>
       <Content
-        currentCamera={currentCamera}
         cameraButtons={cameraButtons}
         propertyId={block?.propertyId}
         isEditable={props.isEditable}
