+++ web/src/beta/features/Visualizer/shared/types.ts
@@ -1,6 +1,5 @@
 import { Theme } from "@reearth/beta/features/Visualizer/Crust/types";
 import { ValueType, ValueTypes } from "@reearth/beta/utils/value";
-import type { Camera } from "@reearth/beta/utils/value";
 import type { Layer, FlyTo } from "@reearth/core";
 
 export type InstallableBlock = {
@@ -27,7 +26,6 @@ export type CommonBlockProps<T = any> = {
   isSelected?: boolean;
   block?: T;
   theme?: Theme;
-  currentCamera?: Camera | undefined;
   padding?: {
     bottom: number;
     top: number;
